feat(refactor-find-inactive-workplaces): Move nextEmail
