+++ server/models/email-campaigns/inactive-workplaces/findInactiveWorkplaces.js
@@ -1,11 +1,11 @@
 const moment = require('moment');
 
 const models = require('../../index');
-const nextEmail = require('./nextEmail');
+const nextEmail = require('../../../services/email-campaigns/inactive-workplaces/nextEmail');
 
 const refreshInactiveWorkplaces = async () => {
   return models.sequelize.query('REFRESH MATERIALIZED VIEW cqc."LastUpdatedEstablishments"');
-}
+};
 
 const getInactiveWorkplaces = async () => {
   const lastMonth = moment().subtract(1, 'months');
@@ -59,7 +59,7 @@ const getInactiveWorkplaces = async () => {
       },
     },
   );
-}
+};
 
 const transformInactiveWorkplaces = (inactiveWorkplace) => {
   const { id, name } = nextEmail.getTemplate(inactiveWorkplace);
@@ -79,15 +79,13 @@ const transformInactiveWorkplaces = (inactiveWorkplace) => {
       email: inactiveWorkplace.PrimaryUserEmail,
     },
   };
-}
+};
 
 const findInactiveWorkplaces = async () => {
   await refreshInactiveWorkplaces();
   const inactiveWorkplaces = await getInactiveWorkplaces();
 
-  return inactiveWorkplaces
-    .filter(nextEmail.shouldReceive)
-    .map(transformInactiveWorkplaces);
+  return inactiveWorkplaces.filter(nextEmail.shouldReceive).map(transformInactiveWorkplaces);
 };
 
 module.exports = {
